requirements.txt
asgiref==3.8.1
attrs==25.1.0
cffi==1.17.1
cryptography==44.0.2
Django==5.1.6
django-cors-headers==4.7.0
django-hashid-field==3.4.1
django-ordered-model==3.7.4
django-secured-fields==0.4.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
hashids==1.3.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
